% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyseSpecimenColours.R,
%   R/AnalyseSpecimenColours_20251013.R
\name{add.colour.models.from.RGB}
\alias{add.colour.models.from.RGB}
\title{Convert to other colour models starting from RGB}
\usage{
add.colour.models.from.RGB(colour.data, R.col, G.col, B.col)

add.colour.models.from.RGB(colour.data, R.col, G.col, B.col)
}
\arguments{
\item{colour.data}{Data frame containing sample information including Accession Number (or any unique identifier for each sample), sample grouping(s), and a column each for R, G and B values}

\item{R.col}{Name of column containing R values}

\item{G.col}{Name of column containing G values}

\item{B.col}{Name of column containing B values}
}
\description{
Convert to other colour models starting from RGB

Convert to other colour models starting from RGB
}
