% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyseSpecimenColours.R,
%   R/AnalyseSpecimenColours_20251013.R
\name{calculate.overlap.for.all.colours}
\alias{calculate.overlap.for.all.colours}
\title{Helper function: do not export. Calculate overlap for individual colours based on all colour models}
\usage{
calculate.overlap.for.all.colours(
  colour.data,
  colour.data.from.all.colour.models,
  accession.number,
  population,
  threshold.1 = 0.75,
  threshold.2 = 0.99
)

calculate.overlap.for.all.colours(
  colour.data,
  colour.data.from.all.colour.models,
  accession.number,
  population,
  threshold.1 = 0.75,
  threshold.2 = 0.99
)
}
\arguments{
\item{colour.data}{Data frame containing sample information including Accession Number (or any unique identifier for each sample), sample grouping(s), and a column each for R, G and B valuesusethis::use_git_config(user.name = "Your Name", user.email = "youremail@example.com")}

\item{colour.data.from.all.colour.models}{Results from add.colour.models.from.RGB()}

\item{accession.number}{Name of column with Accession Number}

\item{population}{Name of column with population assignment (e.g. subspecies, general locality etc.)}

\item{threshold.1}{default value 0.75. Threshold value of the percentage of population that must be separable from threshold.2 of the other population. To test the Patten and Unitt (2002) rule, change this value to 0.95 and keep threshold.2 as 0.99.}

\item{threshold.2}{default value 0.99. Threshold value of the percentage of population that must be separable from threshold.1 of the other population.}
}
\description{
Helper function: do not export. Calculate overlap for individual colours based on all colour models

Helper function: do not export. Calculate overlap for individual colours based on all colour models
}
