% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyseSpecimenColours.R
\name{check.linear.overlap.Amadon}
\alias{check.linear.overlap.Amadon}
\title{Check overlap between pairs based on linear value i.e. 75\% subspecies rule by Amadon (1949). Default values of 0.75 or threshold.1 and 0.99 for threshold 2, change as required.}
\usage{
check.linear.overlap.Amadon(
  data.1,
  data.2,
  name.data.1,
  name.data.2,
  threshold.1 = 0.75,
  threshold.2 = 0.99
)
}
\arguments{
\item{data.1}{Vector values from second dataset}

\item{name.data.1}{Name of first dataset}

\item{name.data.2}{Name of second dataset}

\item{threshold.1}{default value 0.75. Threshold value of the percentage of population that must be separable from threshold.2 of the other population. To test the Patten and Unitt (2002) rule, change this value to 0.95 and keep threshold.2 as 0.99.}

\item{threshold.2}{default value 0.99. Threshold value of the percentage of population that must be separable from threshold.1 of the other population.}
}
\description{
Check overlap between pairs based on linear value i.e. 75\% subspecies rule by Amadon (1949). Default values of 0.75 or threshold.1 and 0.99 for threshold 2, change as required.
}
