% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyseSpecimenColours.R
\name{flatten.list}
\alias{flatten.list}
\title{Helper function: do not export. Flatten nested lists}
\usage{
flatten.list(list.to.flatten, parent.name = "")
}
\arguments{
\item{list.to.flatten}{List with sublists that need to be flattened}

\item{parent.name}{Name to rename flattened lists to}
}
\description{
Helper function: do not export. Flatten nested lists
}
