% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyseSpecimenColours.R,
%   R/AnalyseSpecimenColours_20251013.R
\name{neighbour.joining}
\alias{neighbour.joining}
\title{Create neighbour-joining tree based on RGB input and produce phylogram}
\usage{
neighbour.joining(
  colour.data,
  accession.number,
  population,
  R.col,
  G.col,
  B.col,
  export.figures = FALSE,
  output.folder = NULL,
  output.file.prefix = NULL
)

neighbour.joining(
  colour.data,
  accession.number,
  population,
  R.col,
  G.col,
  B.col,
  export.figures = FALSE,
  output.folder = NULL,
  output.file.prefix = NULL
)
}
\arguments{
\item{colour.data}{Data frame containing sample information including Accession Number (or any unique identifier for each sample), sample grouping(s), and a column each for R, G and B values}

\item{accession.number}{Name of column with Accession Number}

\item{population}{Name of column with population assignment (e.g. subspecies, general locality etc.)}

\item{R.col}{Name of column containing R values}

\item{G.col}{Name of column containing G values}

\item{B.col}{Name of column containing B values}

\item{export.figures}{TRUE/FALSE.  Defaults to FALSE. TRUE exports the phylogram as an image file.}

\item{output.folder}{Path to folder where image file will be exported into. If left blank, the file will be created in present working directory. If a folder with the name designated here doesn't exist in the present workign directory, it will be created.}

\item{output.file.prefix}{Specify file name of phylogram image file if desired. Otherwise the file will be nj_phylogram.jpg}
}
\description{
Create neighbour-joining tree based on RGB input and produce phylogram

Create neighbour-joining tree based on RGB input and produce phylogram
}
