% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyseSpecimenColours.R,
%   R/AnalyseSpecimenColours_20251013.R
\name{pairwise.delta}
\alias{pairwise.delta}
\title{Calculates the CIEDE2000 values between every pair of samples in the dataset.}
\usage{
pairwise.delta(
  colour.data,
  accession.number,
  R.col,
  G.col,
  B.col,
  as.distance.matrix = FALSE
)

pairwise.delta(
  colour.data,
  accession.number,
  R.col,
  G.col,
  B.col,
  as.distance.matrix = FALSE
)
}
\arguments{
\item{colour.data}{Data frame containing sample information including Accession Number (or any unique identifier for each sample), sample grouping(s), and a column each for R, G and B values}

\item{accession.number}{Name of column with Accession Number}

\item{R.col}{Name of column containing R values}

\item{G.col}{Name of column containing G values}

\item{B.col}{Name of column containing B values}

\item{as.distance.matrix}{TRUE/FALSE. Defaults to FALSE. TRUE = outputs pairwise distance as R's distance matrix format. FALSE = outputs pairwise distances as mirrored matrix.}
}
\description{
Calculates the CIEDE2000 values between every pair of samples in the dataset.

Calculates the CIEDE2000 values between every pair of samples in the dataset.
}
