% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyseSpecimenColours.R
\name{pairwise.group.Amadon}
\alias{pairwise.group.Amadon}
\title{Runs check.linear.overlap.Amadon() on all pairwise combinations of groups for one parameter. Default values of 0.75 or threshold.1 and 0.99 for threshold 2, change as required.}
\usage{
pairwise.group.Amadon(
  colour.data,
  accession.number,
  population,
  parameter.to.check.overlap,
  threshold.1 = 0.75,
  threshold.2 = 0.99
)
}
\arguments{
\item{colour.data}{Data frame containing sample information including Accession Number (or any unique identifier for each sample), sample grouping(s), and a column each for R, G and B values}

\item{accession.number}{Name of column with Accession Number}

\item{population}{Name of column with population assignment (e.g. subspecies, general locality etc.)}

\item{parameter.to.check.overlap}{name of column with values to test subspecies rule by Amadon (1949) on}

\item{threshold.1}{default value 0.75. Threshold value of the percentage of population that must be separable from threshold.2 of the other population. To test the Patten and Unitt (2002) rule, change this value to 0.95 and keep threshold.2 as 0.99.}

\item{threshold.2}{default value 0.99. Threshold value of the percentage of population that must be separable from threshold.1 of the other population.}
}
\description{
Runs check.linear.overlap.Amadon() on all pairwise combinations of groups for one parameter. Default values of 0.75 or threshold.1 and 0.99 for threshold 2, change as required.
}
