% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyseSpecimenColours.R,
%   R/AnalyseSpecimenColours_20251013.R
\name{pca.lda.overlap}
\alias{pca.lda.overlap}
\title{Test 75\% subspecies rule after running PCA and LDA on colours. Can be done for all colour models if the input for colour.data.from.all.colour.models is the results of add.colour.models.from.RGB()}
\usage{
pca.lda.overlap(
  colour.data,
  accession.number,
  population,
  R.col,
  G.col,
  B.col,
  threshold.1 = 0.75,
  threshold.2 = 0.99,
  colour.data.from.all.colour.models = NULL,
  export.figures = FALSE,
  output.folder = NULL
)

pca.lda.overlap(
  colour.data,
  accession.number,
  population,
  R.col,
  G.col,
  B.col,
  threshold.1 = 0.75,
  threshold.2 = 0.99,
  colour.data.from.all.colour.models = NULL,
  export.figures = FALSE,
  output.folder = NULL
)
}
\arguments{
\item{colour.data}{Data frame containing sample information including Accession Number (or any unique identifier for each sample), sample grouping(s), and a column each for R, G and B values}

\item{accession.number}{Name of column with Accession Number}

\item{population}{Name of column with population assignment (e.g. subspecies, general locality etc.)}

\item{R.col}{Name of column containing R values}

\item{G.col}{Name of column containing G values}

\item{B.col}{Name of column containing B values}

\item{threshold.1}{default value 0.75. Threshold value of the percentage of population that must be separable from threshold.2 of the other population. To test the Patten and Unitt (2002) rule, change this value to 0.95 and keep threshold.2 as 0.99.}

\item{threshold.2}{default value 0.99. Threshold value of the percentage of population that must be separable from threshold.1 of the other population.}

\item{colour.data.from.all.colour.models}{If running the subspecies rule by Amadon (1949) for other colour models, the input for this will be the results from add.colour.models.from.RGB()}

\item{export.figures}{TRUE/FALSE.  Defaults to FALSE. TRUE exports the PCAs and LDAs as image files.}

\item{output.folder}{Path to folder where image file will be exported into. If left blank, the file will be created in present working directory. If a folder with the name designated here doesn't exist in the present workign directory, it will be created.}
}
\description{
Test 75\% subspecies rule after running PCA and LDA on colours. Can be done for all colour models if the input for colour.data.from.all.colour.models is the results of add.colour.models.from.RGB()

Test 75\% subspecies rule after running PCA and LDA on colours. Can be done for all colour models if the input for colour.data.from.all.colour.models is the results of add.colour.models.from.RGB()
}
