% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyseSpecimenColours.R
\name{run.functions}
\alias{run.functions}
\title{Conduct all analyses described in Teo et al. (2025). Choose between key analyses only or all analyses (i.e. including exploratory analyses)}
\usage{
run.functions(
  colour.data,
  accession.number,
  population,
  R.col,
  G.col,
  B.col,
  run,
  threshold.1 = 0.75,
  threshold.2 = 0.99,
  export.colours.plot = FALSE,
  export.figures = FALSE,
  export.excel = FALSE,
  output.folder = NULL,
  excel.file.name = NULL
)
}
\arguments{
\item{colour.data}{Data frame containing sample information including Accession Number (or any unique identifier for each sample), sample grouping(s), and a column each for R, G and B values}

\item{accession.number}{Name of column with Accession Number}

\item{population}{Name of column with population assignment (e.g. subspecies, general locality etc.)}

\item{R.col}{Name of column containing R values}

\item{G.col}{Name of column containing G values}

\item{B.col}{Name of column containing B values}

\item{run}{"key" or "all". "key" runs the analyses found to be important in Teo et al. (2025), "all" runs analyses including exploratory analyses that were conducted}

\item{threshold.1}{default value 0.75. Threshold value of the percentage of population that must be separable from threshold.2 of the other population. To test the Patten and Unitt (2002) rule, change this value to 0.95 and keep threshold.2 as 0.99.}

\item{threshold.2}{default value 0.99. Threshold value of the percentage of population that must be separable from threshold.1 of the other population.}

\item{export.colours.plot}{TRUE/FALSE. Defaults to FALSE. TRUE exports boxplots of individual colour values (e.g. R, G and B) for all colour models, for all pairwise group comparisons}

\item{export.figures}{TRUE/FALSE. Defaults to FALSE. TRUE exports the LDA and PCA figures into an output file.}

\item{export.excel}{TRUE/FALSE. TRUE exports results of the Amadon (1949) subspecies rule test as an .xlsx file.}

\item{output.folder}{Path to folder where image file will be exported into. If left blank, the file will be created in present working directory. If a folder with the name designated here doesn't exist in the present workign directory, it will be created.}

\item{excel.file.name}{Specify file name of excel file (if export.excel = TRUE). Otherwise the file will be named diagnosability.xlsx}
}
\description{
Conduct all analyses described in Teo et al. (2025). Choose between key analyses only or all analyses (i.e. including exploratory analyses)
}
